Adds checks for and unexports module shell funcs
When these are present and exported in the job environment, they get
passed through qrsh which causes a bunch of noisy (but harmless) error
messages.

Because the environment is passed through anyway, the module functions
should almost never be needed in the environment of something run off
the mpirun worker processes, and if they are, that process can redefine
them itself.

This hopefully should no longer be necessary after a change in
scheduler.